Added support for unix domain sockets and custom parsers.
Added the ability to use tcp sockets or unix domain sockets. Also added the
ability to specify a parser to use when processing responses from the redis
server.  To do this I had to bring back the custom connection pool, since
you need to instantiate a custom connection pool to be able to specify a
parser class.