Fix add() with a negative timeout
Adding to the cache with a negative timeout should be immediately
invalid. So don't bother adding it. The add is considered failed if the
key does not exist.

Fixing this allowed restoring a test from Django's session test suite,
SessionTestsMixin.test_actual_expiry. Using django-redis as a cache with
the Session cache backend is now in full compliance.