Remove version 3.8.x from documentation
This version is clearly no longer supported, as it hasn't been
touched since 2015 and the documentation no longer lines up with
what that version actually supports. At this point, it makes more
sense to remove any references to that version than it does to
keep the documentation up to date with a version from 3 years ago.